requirements/base.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
django-js-asset==1.0.0
django-mptt==0.9.0
djangorestframework==3.8.2
emoji==0.5.0
future==0.16.0
idna==2.6
oauthlib==2.0.7
Pillow==5.1.0
PyJWT==1.6.1
python-telegram-bot==10.1.0
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
